+++ scripts/tests/assets/exist.ts
@@ -1,8 +1,8 @@
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import * as core from '@actions/core';
-import { database } from '../../src/assets/soundDb.ts';
-import { flatten } from '../../scripts/helpers.ts';
+import { database } from '../../../src/assets/soundDb.ts';
+import { flatten } from '../../helpers.ts';
 
 const flattened = flatten(database);
 const existingEntries = Object.values(flattened)
